shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiokafka>=0.12.0",
    "dependency-injector>=4.48.2",
    "mypy>=1.18.2",
    "pedantic>=2.3.1",
    "pyaml>=25.7.0",
    "pydantic>=2.12.4",
]
